+++ packages/dso-toolkit/src/components/info-button/info-button.stories-of.ts
@@ -11,14 +11,15 @@ import { InfoButton } from "./info-button.models.js";
 type InfoButtonStory = StoryObj<InfoButtonArgs, Renderer>;
 
 interface InfoButtonStories {
-  Active: InfoButtonStory;
-  Inactive: InfoButtonStory;
+  Default: InfoButtonStory;
+  Information: InfoButtonStory;
   SecondaryActive: InfoButtonStory;
   SecondaryInactive: InfoButtonStory;
 }
 
 export interface InfoButtonTemplates<TemplateFnReturnType> {
-  infoButtonTemplate: (infoButtonProperties: InfoButton) => TemplateFnReturnType;
+  infoButtonTemplate: (infoButtonProperties: InfoButton<TemplateFnReturnType>) => TemplateFnReturnType;
+  children?: TemplateFnReturnType;
 }
 
 interface InfoButtonStoriesParameters<Implementation, Templates, TemplateFnReturnType> extends StoriesParameters<
@@ -48,28 +49,29 @@ export function infoButtonMeta<TRenderer extends Renderer>({ readme }: MetaOptio
   };
 }
 
+const baseStoryParameters = {
+  parameters: { layout: "centered" },
+};
+
 export function infoButtonStories<Implementation, Templates, TemplateFnReturnType>({
   storyTemplates,
   templateContainer,
 }: InfoButtonStoriesParameters<Implementation, Templates, TemplateFnReturnType>): InfoButtonStories {
   return {
-    Active: {
-      args: {
-        active: true,
-      },
+    Default: {
+      ...baseStoryParameters,
       render: templateContainer.render(storyTemplates, (args, { infoButtonTemplate }) =>
         infoButtonTemplate(infoButtonArgsMapper(args)),
       ),
     },
-    Inactive: {
-      args: {
-        active: false,
-      },
-      render: templateContainer.render(storyTemplates, (args, { infoButtonTemplate }) =>
-        infoButtonTemplate(infoButtonArgsMapper(args)),
+    Information: {
+      ...baseStoryParameters,
+      render: templateContainer.render(storyTemplates, (args, { infoButtonTemplate, children }) =>
+        infoButtonTemplate(infoButtonArgsMapper(args, children)),
       ),
     },
     SecondaryActive: {
+      ...baseStoryParameters,
       args: {
         active: true,
         secondary: true,
@@ -79,6 +81,7 @@ export function infoButtonStories<Implementation, Templates, TemplateFnReturnTyp
       ),
     },
     SecondaryInactive: {
+      ...baseStoryParameters,
       args: {
         active: false,
         secondary: true,