+++ storybook/cypress/e2e/info-button.cy.ts
@@ -1,0 +1,37 @@
+describe("Info Button", () => {
+  const variants = ["default", "information", "secondary-inactive", "secondary-active"];
+
+  variants.forEach((variant) => {
+    it(`${variant} should be accessible and match snapshot`, () => {
+      cy.visit(`http://localhost:45000/iframe.html?id=core-info-button--${variant}`);
+      cy.injectAxe();
+      cy.dsoCheckA11y("dso-info-button.hydrated");
+      cy.get("dso-info-button.hydrated").matchImageSnapshot(`${Cypress.currentTest.title}`);
+    });
+  });
+
+  describe("information toggletip", () => {
+    beforeEach(() => {
+      cy.visit("http://localhost:45000/iframe.html?id=core-info-button--information");
+    });
+
+    it(`should show the tooltip on hover`, () => {
+      cy.get("dso-info-button.hydrated")
+        .realHover()
+        .shadow()
+        .find("dso-icon-button")
+        .shadow()
+        .find(".dso-tooltip")
+        .should("be.visible");
+    });
+
+    it("should show the toggletip on click", () => {
+      cy.get("dso-info-button.hydrated").realClick().shadow().find("> .dso-tooltip").should("have.class", "visible");
+      cy.get("dso-info-button.hydrated")
+        .realClick()
+        .shadow()
+        .find("> .dso-tooltip")
+        .should("not.have.class", "visible");
+    });
+  });
+});